Add files via upload
This repository contains a collection of local language lexicons and data related to sentiment analysis. The dataset is designed to facilitate sentiment analysis tasks and provide resources for researchers and developers working with local languages. 

The repository includes the following components:
1. Local language lexicons: A comprehensive set of lexicons for local languages, organized in separate files.
2. Sentiment analysis data: Labeled datasets for sentiment analysis tasks in local languages, with annotations indicating positive, negative, or neutral sentiment.

Please note that this data is made available for research and academic purposes. If you utilize this dataset, kindly acknowledge the source by citing the appropriate references (if any provided) or by mentioning this GitHub repository in your work. Feel free to explore the data and utilize it in your research or application development. If you have any questions, suggestions, or improvements, please don't hesitate to open an issue or reach out to us. We hope that this repository will contribute to the advancement of local language sentiment analysis research and foster collaboration among the research community.